+++ macros/generic_tests/not_null_proportion.sql
@@ -10,53 +10,7 @@
 
 with validation as (
   select
-    sum(case when {{ column_name }} is null then 0 else 1 end) / cast(count(*) as numeric) as not_null_proportion
-  from {{ model }}
-),
-validation_errors as (
-  select
-    not_null_proportion
-  from validation
-  where not_null_proportion < {{ at_least }} or not_null_proportion > {{ at_most }}
-)
-select
-  *
-from validation_errors
-
-{% endmacro %}
-
-{% macro athena__test_not_null_proportion(model) %}
-
-{% set column_name = kwargs.get('column_name', kwargs.get('arg')) %}
-{% set at_least = kwargs.get('at_least', kwargs.get('arg')) %}
-{% set at_most = kwargs.get('at_most', kwargs.get('arg', 1)) %}
-
-with validation as (
-  select
-    sum(case when {{ column_name }} is null then 0 else 1 end) / cast(count(*) as double) as not_null_proportion
-  from {{ model }}
-),
-validation_errors as (
-  select
-    not_null_proportion
-  from validation
-  where not_null_proportion < {{ at_least }} or not_null_proportion > {{ at_most }}
-)
-select
-  *
-from validation_errors
-
-{% endmacro %}
-
-{% macro presto__test_not_null_proportion(model) %}
-
-{% set column_name = kwargs.get('column_name', kwargs.get('arg')) %}
-{% set at_least = kwargs.get('at_least', kwargs.get('arg')) %}
-{% set at_most = kwargs.get('at_most', kwargs.get('arg', 1)) %}
-
-with validation as (
-  select
-    sum(case when {{ column_name }} is null then 0 else 1 end) / cast(count(*) as double) as not_null_proportion
+    sum(case when {{ column_name }} is null then 0 else 1 end) / {{ dbt_utils.safe_cast('count(*)', dbt_utils.type_numeric() ) }} as not_null_proportion
   from {{ model }}
 ),
 validation_errors as (